Change the way dependancies are built/used
Instead of deleting the group entirely every time now, it'll simply only
remove the fieldset from the DOM, causing the fields not to be
submitted. However, doing it this way means user input is saved, in case
the dependancy is triggered again (think accidental clicks on radio
buttons or whatever).

Also made sure the received value is always an array, in case of
checkboxes/multiple selects.